🛠️ src/extension.ts -> Modified imports, updated code to use new DocuGen class 🔴 src/logic.ts 🛠️ src/constants.ts -> Moved to src/logic/constants.ts, removed VSCode imports 🟢 src/logic/docugen.ts 🟢 src/logic/index.ts 🛠️ src/interfaces/FileSection.ts -> Moved to src/logic/interfaces/FileSection.ts 🟢 src/logic/interfaces/IDocuGen.ts 🟢 src/logic/interfaces/IModelProvider.ts 🟢 src/logic/interfaces/ISecretProvider.ts 🟢 src/logic/package.json 🟢 src/logic/providers.ts 🛠️ src/test/extension.test.ts -> Moved to src/logic/test/extension.test.ts 🔴 src/providers.ts
