+++ src/logic/index.ts
@@ -1,0 +1,1 @@
+export { DocuGen } from './docugen';
