+++ src/logic/interfaces/FileSection.ts
