+++ src/logic/interfaces/IDocuGen.ts
@@ -1,0 +1,3 @@
+export interface IDocuGen{
+    scanRepository(workspaceFsPath: string, excludeItemsFilePaths: string[], excludeExtensionsFilePaths: string[], itemsToBeIncludedFilePaths: (string | undefined)[], documentFilePath: string): void;
+}
