+++ src/logic/interfaces/IModelProvider.ts
@@ -1,0 +1,3 @@
+export interface IModelProvider {
+    sendRequestToModel(prompt: string, content: string, fileName: string): Promise<any>;
+}
