+++ src/logic/interfaces/ISecretProvider.ts
@@ -1,0 +1,4 @@
+// ISecretProvider.ts
+export interface ISecretProvider {
+    getSecret(key: string): Promise<string | undefined>;
+}
